Option to integrate with pluralkit, which disables forwarding messages that will be proxied
Pluralkit generally gets used by the user setting a prefix/suffix, and sending messages
with it. The bot then deletes the user's original message and replaces it with one with
the desired username, avatar, etc.
When bridging from discord to a platform that doesn't support message deletions (eg IRC),
this causes most pluralkit messages to be sent twice.

This adds a new option, `AllowPluralKit`, which when enabled checks if the pluralkit bot
is present in each guild. If it is, then we will attempt to query each user's members list,
make a regex for the messages that will be proxied by pluralkit, and then don't forward
messages from that user matching that regex (waiting for the bot to send it instead).

This isn't perfect - it requires each discord user to have their members list public, and
doesn't support autoproxying (since there's no way to get those settings from the pluralkit API).
But it's better than nothing.

Closes #2153